backend/requirements.txt
streamlit~=1.23.1
openai~=0.27.7
fastapi~=0.96.0
pydantic~=1.10.9
uvicorn~=0.22.0
ffmpeg-python~=0.2.0
python-multipart~=0.0.6
git+https://github.com/aarnphm/whispercpp.git